pyproject.toml
[tool.poetry]
name = "voice-inline-bot"
version = "0.1.0"
description = ""
authors = ["mircuss <mircusss69@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.7.0"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
